requirements.txt
google-api-python-client==2.108.0
python-dotenv==1.0.0
streamlit==1.32.0
google-generativeai==0.3.2 